Refacto tests and mock for titelive_stocks
